I implement fun 'searchItemFromCategory' on repository pattern
